Say double where I mean 64-bit, not float
FLOAT and DOUBLE are the same 64-bit type on Snowflake, so every one of
these 30 casts means exactly what it meant before and no value moves.
They are not the same type everywhere else -- DuckDB's FLOAT is 32-bit,
and a cast through it turns 12345.6789 into 12345.6787109375.

That matters most in stable_sum, whose whole job is to protect the
precision of an exact-decimal sum; handing the result to a 32-bit cast
would throw the guarantee away at the last step. The MLB-9 spike caught
this live in the value A/B -- 553.5999755859375 against prod's 553.7 --
on precisely the lineage that still said `as float`.

Snowflake keeps the seed's scoring weights 64-bit either way; the seed
type changes for the same reason the casts do.

+++ dbt_league/models/intermediate/int_cbs__player_daily.sql
@@ -276,9 +276,9 @@ select
         else d.lineup_slot != 'RS'
     end                                          as is_active_slot,
     d.games_played,
-    cast(null as float)                          as platform_points,
-    cast(null as float)                          as platform_hitting_pts,
-    cast(null as float)                          as platform_pitching_pts,
+    cast(null as double)                          as platform_points,
+    cast(null as double)                          as platform_hitting_pts,
+    cast(null as double)                          as platform_pitching_pts,
 
     -- Hitting counting stats (CBS scores R/RBI/BB/SB/TB; H/AB/HBP/SF +
     -- HR/2B/3B ride as unpriced context -- the last three feed the
